requirements.txt
openai
streamlit
requests
beautifulsoup4
google-cloud-storage